cli/self: Stop printing information about the node
If we ever want to disentangle the users' identity from the node, then
the `rad self` command should not display information related to the
node so prominently.

The relation between `rad` and `radicle-node` is really similar to that
between `rad` and SSH Agent. They communicate via socket. So, when this
connection is successful, we print that, but not more.

Users might have to re-learn a tiny bit here, but I wager it is worth
the pain.

Note that the information being removed here is available via
`rad node status` (see previous commit).