node: Remove dependency on `anyhow`
It has always bugged me that we resort to `anyhow` for just a handful of
errors, really. It also turns out that we were holding `lexopt` wrong,
it features `parse_with` which neatly integrates with `FromStr`, and
luckily all our arguments implement that trait. This makes for cleaner
option parsing.

For now, the execution errors are all "transparent", which shouldn't be
much of a regression, if at all.