remote-helper: Update Patch by ID Prefix
`rad patch list` does not show full patch IDs by default, and overly
long patch IDs are cumbersome to handle.

If a unique prefix is used, allow updating patches identified by
that prefix.

E.g. when there is exactly one patch with an ID that has a prefix of
"a", then pushing to `refs/heads/patches/a` updates that patch.