REVIEW
First thing was to add an `args` module.

I then duplicated the `ScopeParser` code to get the same functionality. I chose
to duplicate it for the time being and clean it up later.

It's important to note that defaulting `public` subtly changed the behaviour of the evaluation.
If neither were set, we would always end up with a `Visibility::Public`, which would mean we never hit the interactive selection later.
This is fixed by having an `Args::visibility` method to return an `Option<Visibility>` and the calling code works as per usual.

To clean up the calling code, I also introudced `Args::interactive` and `Args::seed`.