node: only consider onion addresses if onion is configured
There is an unfortunate lack of single-responsibility with handling the
addresses in our code at the moment.
This means that adding addresses to the table, and reading them from the table
are spread across the code base.

A couple of places were found to further filter out onion addresses, if the
configuration does not have an onion address.

This is a band-aid fix until the point in time where this can better improved.