Revert "node: Use `std::time` for reactor and wire"
This reverts commit e404f1038f461264f9395742ef74f5b710bef54a.

The commit above introduced the `Epoch` type to reduce the need for
depending on the `localtime` crate. Eventually, node operators noticed
that announcements would appear to be in the future and cause
connection issues.

After more research, an issue[^1] was found that describes how on some
platforms, primarily Linux and MacOS, `Instant::duration_since` would
return a shortened duration. This is the cause of this apparent drift
in time.

[^1]: https://github.com/rust-lang/rust/issues/87906