protocol/connections: testing Connections
Adds property based tests to ensure the correctness of the
`Connections` logic.

The basis of the tests is to ensure that the state transitions are
correct, and that the machine cannot end up in an unexpected state.

There are also tests that ensure that certain state is preserved
across transitions, as well as other domain logic.

Disclaimer: the basis of the test suite was generated using Claude
with human input on what properties to start off with. All generated
code was reviewed for correctness and modified to fit the style of the
project as best as possible.

Squash to tests