protocol/connections: refactor config
Refactor the configuration of connections.

This places the config types and values into their own sub-module.
It then refactors to have more logical units for sub-configuration:
durations, inbound, and outbound.

It also defines the constants that can be used as the default values,
which will replace the constants defined in `service`.